Created a config package and moved the config class to it
