+++ src/oops/SOLID/singleResponsibilityPrinciple/before/config/EmployeeConfig.java
@@ -1,4 +1,4 @@
-package oops.SOLID.singleResponsibilityPrinciple.before;
+package oops.SOLID.singleResponsibilityPrinciple.before.config;
 
 import oops.SOLID.singleResponsibilityPrinciple.before.model.Employee;
